requirements/dev.txt
# Fast API
uvicorn
fastapi[all]==0.111.0
python-multipart==0.0.9
# Database
sqlalchemy==2.0.31
# Cryptography and Security
passlib==1.7.4
cryptography==42.0.8
# Environment
python-dotenv==1.0.1
# JWT
python-jose==3.3.0

# mysql
mysql-connector-python==8.4.0



## google drive
google-api-python-client==2.135.0 
google-auth-oauthlib==1.2.0 
google-auth-httplib2==0.2.0 
google-auth==2.31.0 

# CNN
tensorflow
tensorflow-cpu
h5py==3.11.0
opencv-python==4.10.0.84
numpy<2.0.0


# Face recognition

face_recognition==1.3.0

# Gemini chatbot

langchain_google_genai==1.0.7